% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{qanova}
\alias{qanova}
\title{ANOVA with FO, PQ, and TWI terms.}
\usage{
qanova(formula, data, vars)
}
\arguments{
\item{formula}{Complete formula for the model.}

\item{data}{Data frame.}

\item{vars}{Character vector of variable names for ANOVA.}

\item{formula}{Complete formula for the model.}
}
\value{
A data frame with ANOVA statistics.

The ANOVA is not fully sequential. The FO and PQ terms are
added for the first variable, followed by the TWI terms.
The FO and PQ terms for the second variable are added only
to the FO and PQ model for the first variable since there
is overlap in the TWI terms. Thus the FO and PQ terms for
the final variable are added to all other FO and PQ terms
but no TWI terms.
}
\description{
ANOVA with FO, PQ, and TWI terms.
}
